# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ASTRAL-BOT Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guilds/{guildId}/config:
    get:
      operationId: getGuildConfig
      tags: [guilds]
      summary: Get guild configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildConfig"
    put:
      operationId: updateGuildConfig
      tags: [guilds]
      summary: Update guild configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuildConfigBody"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildConfig"
  /guilds/{guildId}/stats:
    get:
      operationId: getGuildStats
      tags: [guilds]
      summary: Get guild statistics
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildStats"
  /guilds/{guildId}/channels:
    get:
      operationId: getGuildChannels
      tags: [guilds]
      summary: Get guild channels
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscordChannel"
  /guilds/{guildId}/warns:
    get:
      operationId: getGuildWarns
      tags: [warns]
      summary: Get all warns for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of warns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warn"
  /guilds/{guildId}/banned-words:
    get:
      operationId: getBannedWords
      tags: [guilds]
      summary: Get banned words for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of banned words
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BannedWord"
    post:
      operationId: addBannedWord
      tags: [guilds]
      summary: Add a banned word
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBannedWordBody"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BannedWord"
  /guilds/{guildId}/banned-words/{wordId}:
    delete:
      operationId: deleteBannedWord
      tags: [guilds]
      summary: Delete a banned word
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: wordId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /guilds/{guildId}/giveaways:
    get:
      operationId: getGiveaways
      tags: [giveaways]
      summary: Get all giveaways for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of giveaways
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Giveaway"
    post:
      operationId: createGiveaway
      tags: [giveaways]
      summary: Create a giveaway
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGiveawayBody"
      responses:
        "201":
          description: Created giveaway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Giveaway"
  /guilds/{guildId}/giveaways/{giveawayId}/end:
    post:
      operationId: endGiveaway
      tags: [giveaways]
      summary: End a giveaway and pick winners
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: giveawayId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Giveaway ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /guilds/{guildId}/surveys:
    get:
      operationId: getSurveys
      tags: [surveys]
      summary: Get all surveys for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of surveys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Survey"
    post:
      operationId: createSurvey
      tags: [surveys]
      summary: Create a survey
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyBody"
      responses:
        "201":
          description: Created survey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
  /guilds/{guildId}/surveys/{surveyId}:
    get:
      operationId: getSurveyById
      tags: [surveys]
      summary: Get a survey with responses
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: surveyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Survey with responses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyWithResponses"
  /guilds/{guildId}/actions/maintenance:
    post:
      operationId: triggerMaintenance
      tags: [actions]
      summary: Toggle maintenance mode
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceBody"
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /guilds/{guildId}/actions/breach:
    post:
      operationId: triggerBreach
      tags: [actions]
      summary: Toggle breach/lockdown mode
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreachBody"
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /guilds/{guildId}/actions/embed:
    post:
      operationId: sendEmbed
      tags: [actions]
      summary: Send an embed to a channel
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmbedBody"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /guilds/{guildId}/actions/announce:
    post:
      operationId: sendAnnouncement
      tags: [actions]
      summary: Send an announcement
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnounceBody"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /guilds/{guildId}/actions/say:
    post:
      operationId: sendSay
      tags: [actions]
      summary: Send a plain message as the bot
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SayBody"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /guilds/{guildId}/logs:
    get:
      operationId: getGuildLogs
      tags: [guilds]
      summary: Get recent logs for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
components:
  schemas:
...
    GuildConfig:
      type: object
      properties:
        guildId:
          type: string
        logChannelId:
          type: string
          nullable: true
        surveyChannelId:
          type: string
          nullable: true
        maintenanceMode:
          type: boolean
        maintenanceReason:
          type: string
          nullable: true
        breachMode:
          type: boolean
        automodEnabled:
          type: boolean
        antiRaidEnabled:
          type: boolean
        antiSpamEnabled:
          type: boolean
        maxMentions:
          type: integer
        maxMessagesPerMinute:
          type: integer
      required: [guildId, maintenanceMode, breachMode, automodEnabled, antiRaidEnabled, antiSpamEnabled, maxMentions, maxMessagesPerMinute]
    UpdateGuildConfigBody:
      type: object
      properties:
        logChannelId:
          type: string
          nullable: true
        surveyChannelId:
          type: string
          nullable: true
        maintenanceMode:
          type: boolean
        maintenanceReason:
          type: string
          nullable: true
        breachMode:
          type: boolean
        automodEnabled:
          type: boolean
        antiRaidEnabled:
          type: boolean
        antiSpamEnabled:
          type: boolean
        maxMentions:
          type: integer
        maxMessagesPerMinute:
          type: integer
    GuildStats:
      type: object
      properties:
        memberCount:
          type: integer
        channelCount:
          type: integer
        roleCount:
          type: integer
        warnCount:
          type: integer
        giveawayCount:
          type: integer
        surveyCount:
          type: integer
        bannedWordCount:
          type: integer
        boostLevel:
          type: integer
        boostCount:
          type: integer
      required: [memberCount, channelCount, roleCount, warnCount, giveawayCount, surveyCount, bannedWordCount, boostLevel, boostCount]
    DiscordChannel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: integer
      required: [id, name, type]
    Warn:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        userId:
          type: string
        username:
          type: string
        reason:
          type: string
        moderatorId:
          type: string
        moderatorName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, guildId, userId, username, reason, moderatorId, moderatorName, createdAt]
    BannedWord:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        word:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, guildId, word, createdAt]
    AddBannedWordBody:
      type: object
      properties:
        word:
          type: string
      required: [word]
    Giveaway:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        channelId:
          type: string
        messageId:
          type: string
          nullable: true
        prize:
          type: string
        winnersCount:
          type: integer
        endsAt:
          type: string
          format: date-time
        ended:
          type: boolean
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        participants:
          type: integer
      required: [id, guildId, channelId, prize, winnersCount, endsAt, ended, createdBy, createdAt, participants]
    CreateGiveawayBody:
      type: object
      properties:
        channelId:
          type: string
        prize:
          type: string
        winnersCount:
          type: integer
        durationMinutes:
          type: integer
        createdBy:
          type: string
      required: [channelId, prize, winnersCount, durationMinutes, createdBy]
    Survey:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        channelId:
          type: string
        responseChannelId:
          type: string
          nullable: true
        questions:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        responseCount:
          type: integer
      required: [id, guildId, title, channelId, questions, active, createdAt, responseCount]
    SurveyWithResponses:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        channelId:
          type: string
        responseChannelId:
          type: string
          nullable: true
        questions:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        responses:
          type: array
          items:
            $ref: "#/components/schemas/SurveyResponse"
      required: [id, guildId, title, channelId, questions, active, createdAt, responses]
    SurveyResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        answers:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, username, answers, createdAt]
    CreateSurveyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        channelId:
          type: string
        responseChannelId:
          type: string
          nullable: true
        questions:
          type: array
          items:
            type: string
      required: [title, channelId, questions]
    MaintenanceBody:
      type: object
      properties:
        enabled:
          type: boolean
        reason:
          type: string
          nullable: true
      required: [enabled]
    BreachBody:
      type: object
      properties:
        enabled:
          type: boolean
        reason:
          type: string
          nullable: true
      required: [enabled]
    SendEmbedBody:
      type: object
      properties:
        channelId:
          type: string
        title:
          type: string
        description:
          type: string
        color:
          type: string
          nullable: true
        footer:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
      required: [channelId, title, description]
    AnnounceBody:
      type: object
      properties:
        channelId:
          type: string
        title:
          type: string
        message:
          type: string
        pingEveryone:
          type: boolean
      required: [channelId, title, message, pingEveryone]
    SayBody:
      type: object
      properties:
        channelId:
          type: string
        message:
          type: string
      required: [channelId, message]
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        action:
          type: string
        targetId:
          type: string
          nullable: true
        targetName:
          type: string
          nullable: true
        moderatorId:
          type: string
          nullable: true
        moderatorName:
          type: string
          nullable: true
        details:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, guildId, action, createdAt]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]