# lib/api-zod/src/generated/types/guildStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ASTRAL-BOT Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface GuildStats {
  memberCount: number;
  channelCount: number;
  roleCount: number;
  warnCount: number;
  giveawayCount: number;
  surveyCount: number;
  bannedWordCount: number;
  boostLevel: number;
  boostCount: number;
}