# lib/api-zod/src/generated/types/updateGuildConfigBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ASTRAL-BOT Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateGuildConfigBody {
  logChannelId?: string | null;
  surveyChannelId?: string | null;
  maintenanceMode?: boolean;
  maintenanceReason?: string | null;
  breachMode?: boolean;
  automodEnabled?: boolean;
  antiRaidEnabled?: boolean;
  antiSpamEnabled?: boolean;
  maxMentions?: number;
  maxMessagesPerMinute?: number;
}